+++ app/src/main/java/com/example/explorerapplication/MyApplication.kt
@@ -1,0 +1,7 @@
+package com.example.explorerapplication
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MyApplication : Application()
