+++ tmdbnavigation/src/main/AndroidManifest.xml
@@ -4,7 +4,7 @@
 
     <application>
         <activity
-            android:name=".ui.view.MoviesActivity"
+            android:name=".ui.views.MoviesActivity"
             android:exported="false" />
     </application>
