+++ tmdbnavigation/src/main/java/com/example/tmdbnavigation/di/MovieModule.kt
@@ -1,0 +1,16 @@
+package com.example.tmdbnavigation.di
+
+import com.example.tmdbnavigation.data.MovieRepositoryImpl
+import com.example.tmdbnavigation.domain.MovieRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class MovieModule {
+
+    @Binds
+    abstract fun bindMovieRepository(movieRepositoryImpl: MovieRepositoryImpl): MovieRepository
+}
