+++ tmdbnavigation/src/main/java/com/example/tmdbnavigation/domain/MovieRepository.kt
@@ -1,5 +1,7 @@
 package com.example.tmdbnavigation.domain
 
+import com.example.tmdbnavigation.domain.models.Movie
+
 interface MovieRepository {
     suspend fun getMovies(): List<Movie>
 }
