+++ tmdbnavigation/src/main/java/com/example/tmdbnavigation/domain/models/Movie.kt
@@ -1,4 +1,4 @@
-package com.example.tmdbnavigation.domain
+package com.example.tmdbnavigation.domain.models
 
 data class Movie(
     val name: String,
