+++ tmdbnavigation/src/main/java/com/example/tmdbnavigation/ui/views/MovieDetailFragment.kt
@@ -1,0 +1,21 @@
+package com.example.tmdbnavigation.ui.views
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.tmdbnavigation.R
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MovieDetailFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_movie_detail, container, false)
+    }
+}
