+++ tmdbnavigation/src/main/java/com/example/tmdbnavigation/ui/views/MoviesActivity.kt
@@ -1,9 +1,11 @@
-package com.example.tmdbnavigation.ui.view
+package com.example.tmdbnavigation.ui.views
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 import com.example.tmdbnavigation.R
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class MoviesActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
