+++ tmdbnavigation/src/main/res/layout/activity_movie_list.xml
@@ -8,4 +8,4 @@
     android:layout_height="match_parent"
     app:defaultNavHost="true"
     app:navGraph="@navigation/movie_nav"
-    tools:context=".ui.view.MoviesActivity" />
+    tools:context=".ui.views.MoviesActivity" />
