+++ tmdbnavigation/src/main/res/layout/fragment_movie_list.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.view.MovieListFragment">
+    tools:context=".ui.views.MovieListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:layout_width="match_parent"
