+++ tmdbnavigation/src/main/res/navigation/movie_nav.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/movieListFragment"
-        android:name="com.example.tmdbnavigation.ui.view.MovieListFragment"
+        android:name="com.example.tmdbnavigation.ui.views.MovieListFragment"
         android:label="movie_list_fragment"
         tools:layout="@layout/fragment_movie_list" >
         <action
